+++ core/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.0.4
 
+- Add timeouts to RPC subscriptions
 - Integrate upstream `rust-libp2p` `0.54` changes to the bootstrap process
 
 ## 1.0.3